+++ vite.config.ts
@@ -19,7 +19,6 @@ export default defineConfig({
     rollupOptions: {
       input: {
         main: path.resolve(__dirname, "index.html"),
-        screenshot: path.resolve(__dirname, "screenshot.html"),
       },
     },
   },